Add job management components and types
- Introduced a modular job management system with dedicated components for job listings, filters, and headers.
- Created reusable components: `JobCard`, `JobList`, `JobFilters`, and `JobHeader` to enhance UI and maintainability.
- Implemented custom hooks `useJobs` and `useJobFilters` for managing job data and filter state.
- Defined types for job items and filters to improve type safety and clarity in the codebase.
- Added a new `CompanyLogo` component for displaying company branding in job listings.
- Refactored the job page to utilize the new components and hooks, improving overall structure and readability.
- Enhanced loading states and error handling for job fetching functionality.
